rename folder and fix case sensitive
